Use common decorator
